file rename & move files to their own transformer dir
