+++ src/v3/src/transformer/button/index.ts
@@ -10,4 +10,4 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-export * from './transformer';
+export * from './transform';
